Check permission for DELETE method

+++ app/api_1_0/smstools.py
@@ -37,21 +37,20 @@ def list_some_sms(kind):
 
     return jsonify(result)
 
-
 def delete_some_sms(kind, message_id):
     if kind not in current_app.config['KINDS']:
         return not_found(None)
-    try:
-        os.remove(current_app.config[kind.upper()] + "/" + message_id)
 
-    except OSError:
-        return not_found(None)
+    if 'ADMIN_ACCOUNTS' in current_app.config and auth.username() in current_app.config['ADMIN_ACCOUNTS']:
+        result = {}
+        try
+            os.remove(current_app.config[kind.upper()] + "/" + message_id)
+            result['deleted'] = kind + '/' + message_id
+            return jsonify(result)
+        except OSError:
+            return not_found(None)
 
-    result = {}
-    result['deleted'] = kind + '/' + message_id
-
-    return jsonify(result)
-
+    return forbidden(None)
 
 def get_some_sms(kind, message_id):
     if kind not in current_app.config['KINDS']:
